move Kafka producer to kafka package
